Refactor model runners and utils
