+++ bink/__init__.py
@@ -1,4 +1,5 @@
 from . import metrics
 from . import callbacks
-from . import modelrunners
+from . import runners
 from .bink import *
+from . import utils
