+++ bink/bink.py
@@ -1,6 +1,6 @@
 import torch
 from torch.utils.data import DataLoader, TensorDataset
-from utils import get_train_valid_sets
+from bink.utils import get_train_valid_sets
 from bink.callbacks.callbacks import CallbackList
 from bink.callbacks.printer import Tqdm
 from torch.autograd import Variable
